database and da user

+++ app/lib/isp_config/database.rb
@@ -10,10 +10,10 @@ module IspConfig
 
     def all
       response = query({
-                         endpoint: '/json.php?sites_database_get',
+                         endpoint: '/json.php?sites_database_get_all_by_user',
                          method: :GET,
                          body: {
-                           primary_id: "-1"
+                          client_id: user.isp_config_id
                          }
                        })
 
@@ -34,7 +34,6 @@ module IspConfig
                           body: database_hash
                         })
 
-      Rails.logger.debug { response.inspect }
       user.isp_databases.create({ isp_config_database_id: response["response"] }) if response.code == "ok"
 
       formatted_response(response, 'create')
@@ -55,22 +54,51 @@ module IspConfig
                           body: body_params
                         })
 
-      Rails.logger.debug { response.inspect }
       formatted_response(response, 'create')
     end
 
-    def destroy(primary_id)
+    def destroy(primary_ids)
+      ## destroy db user
+      db_user_delete_response = destroy_database_user(primary_ids[:db_user_id])
+      return { success: false, message: I18n.t('isp_config.something_went_wrong', message: db_user_delete_response[:message]),response: db_user_delete_response } unless db_user_delete_response[:success]
+      
       response = query({
                          endpoint: '/json.php?sites_database_delete',
                          method: :DELETE,
                          body: {
                            client_id: user.isp_config_id,
-                           primary_id: primary_id
+                           primary_id: primary_ids[:id]
                          }
                        })
 
-      user.ftp_users.find_by_isp_config_database_id(primary_id).destroy if response.code == "ok"
+      user.isp_databases.find_by_isp_config_database_id(primary_ids[:id]).destroy if response.code == "ok"
+
       formatted_response(response, 'delete')
+    end
+
+    def destroy_database_user(db_user_id)
+      response = query({
+                          endpoint: '/json.php?sites_database_user_delete',
+                          method: :DELETE,
+                          body: {
+                            client_id: user.isp_config_id,
+                            primary_id: db_user_id
+                          }
+                        })
+
+      formatted_response(response, 'delete')
+    end
+
+    def find(id)
+      response = query({
+                         endpoint: '/json.php?sites_database_get',
+                         method: :GET,
+                         body: {
+                           primary_id: id
+                         }
+                       })
+
+      formatted_response(response, 'find')
     end
